use windows-specific CREATE_NEW_PROCESS_GROUP flag in tasker.Taskify
